Maps now define all the objects. Logic separation.
Shadow-casting static box2d rectangles can now be defined in the map
using a TMX object layer. Lights are a bit messier, but still can
be defined in the map.

Logic for launching the game, setting the screen and actually loading
and rendering the map has been separated into 3 files to allow
reuse of the structure in a more complete game.